ProgressEvents/tests/submissions/Ms2ger/other.html

<!doctype html>
<title>ProgressEvent remaining tests</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(function() {
  var ev = new ProgressEvent("test")
  var ev2 = document.createEvent("ProgressEvent")
  ev2.initEvent("test", false, false)

  assert_equals(ev.type, ev2.type)
  assert_equals(ev.target, ev2.target)
  assert_equals(ev.currentTarget, ev2.currentTarget)
  assert_equals(ev.eventPhase, ev2.eventPhase)
  assert_equals(ev.bubbles, ev2.bubbles)
  assert_equals(ev.cancelable, ev2.cancelable)
  assert_equals(ev.defaultPrevented, ev.defaultPrevented)
  assert_equals(ev.isTrusted, ev2.isTrusted)
  assert_true(ev.timeStamp >= ev2.timeStamp)

  assert_equals(ev.lengthComputable, ev2.lengthComputable)
  assert_equals(ev.loaded, ev2.loaded)
  assert_equals(ev.total, ev2.total)
})
</script>
